Change parameters of Flaky test analysis plugin

+++ test/tests5.cpp
@@ -9,7 +9,7 @@ using namespace testing;
 TEST(flaky_tests_5, check_greater_than_flaky){
     // Seed the random number generator with the current time
     srand(static_cast<unsigned int>(time(nullptr))+1);
-    int result = rand()/RAND_MAX;
+    double result = rand()/RAND_MAX;
     std::cout<<"Check greater than 0.5 (flaky): number "<<result<<std::endl;
     ASSERT_GT(result, 0.5);
 }
